Post list shows more information Fix: type for timestamp

+++ app/src/main/java/com/ignaeche/reddittopposts/model/PostData.kt
@@ -8,7 +8,7 @@ data class PostData(@PrimaryKey
                     val id: String,
                     val title: String,
                     val author: String,
-                    val created_utc: Int,
+                    val created_utc: Long,
                     val thumbnail: String,
                     val num_comments: Int,
                     val unread: Boolean)
